requirements.txt
fastapi
uvicorn[standard]
numpy
litellm
pydantic
python-dotenv
